refactor: revert moving of MarkerCollection
