+++ junifer/markers/__init__.py
@@ -5,4 +5,5 @@
 # License: AGPL
 
 from .base import BaseMarker
+from .collection import MarkerCollection
 from .parcel import ParcelAggregation
