+++ junifer/pipeline/__init__.py
@@ -3,5 +3,4 @@
 # Authors: Synchon Mandal <s.mandal@fz-juelich.de>
 # License: AGPL
 
-from .collection import MarkerCollection
 from .pipeline_mixin import PipelineStepMixin
